Code simplification
