+++ solana/tests/utils/tbr-wrapper.ts
@@ -1,0 +1,202 @@
+import anchor from '@coral-xyz/anchor';
+import { Keypair, PublicKey, TransactionSignature } from '@solana/web3.js';
+import { Chain } from '@wormhole-foundation/sdk-base';
+import { UniversalAddress } from '@wormhole-foundation/sdk-definitions';
+import {
+  SolanaPriceOracle,
+  SolanaTokenBridgeRelayer,
+  TransferParameters,
+  VaaMessage,
+} from '@xlabs-xyz/solana-arbitrary-token-transfers';
+import { sendAndConfirmIxs, TestProvider, TestsHelper } from './helpers.js';
+
+import testProgramKeypair from '../../programs/token-bridge-relayer/test-program-keypair.json' with { type: 'json' };
+
+const $ = new TestsHelper();
+
+export class TbrWrapper {
+  readonly client: SolanaTokenBridgeRelayer;
+  readonly provider: TestProvider;
+  readonly logs: { [key: string]: string[] };
+  readonly logsSubscriptionId: number;
+
+  constructor(provider: TestProvider, tbrClient: SolanaTokenBridgeRelayer) {
+    this.provider = provider;
+    this.client = tbrClient;
+    this.logs = {};
+
+    this.logsSubscriptionId = provider.connection.onLogs(
+      'all',
+      (l) => (this.logs[l.signature] = l.logs),
+    );
+  }
+
+  static from(
+    provider: TestProvider,
+    accountType: 'owner' | 'admin' | 'regular',
+    oracleClient: SolanaPriceOracle,
+    debug: boolean,
+  ) {
+    const clientProvider =
+      accountType === 'regular' ? provider : { connection: provider.connection };
+
+    const client = new SolanaTokenBridgeRelayer(
+      clientProvider,
+      'Localnet',
+      $.pubkey.from(testProgramKeypair),
+      oracleClient,
+      debug,
+    );
+
+    return new TbrWrapper(provider, client);
+  }
+
+  static async create(provider: TestProvider, debug: boolean) {
+    const client = await SolanaTokenBridgeRelayer.create(
+      { connection: provider.connection },
+      debug,
+    );
+
+    return new TbrWrapper(provider, client);
+  }
+
+  get publicKey(): PublicKey {
+    return this.provider.publicKey;
+  }
+
+  get account() {
+    return this.client.account;
+  }
+
+  get read() {
+    return this.client.read;
+  }
+
+  /** Unregister the logs event so that the test does not hang. */
+  async close() {
+    await this.provider.connection.removeOnLogsListener(this.logsSubscriptionId);
+  }
+
+  displayLogs(signature: string) {
+    let lines = this.logs[signature];
+    if (lines === undefined) {
+      lines = ['<logs not found>'];
+    }
+    console.log(`Signature '${signature}':`);
+    for (const line of lines) {
+      console.log(`  > ${line}`);
+    }
+  }
+
+  async initialize(args: {
+    owner: PublicKey;
+    feeRecipient: PublicKey;
+    admins: PublicKey[];
+  }): Promise<TransactionSignature> {
+    return sendAndConfirmIxs(this.provider, await this.client.initialize(args));
+  }
+
+  async submitOwnerTransferRequest(newOwner: PublicKey): Promise<TransactionSignature> {
+    return sendAndConfirmIxs(this.provider, await this.client.submitOwnerTransferRequest(newOwner));
+  }
+
+  async confirmOwnerTransferRequest(): Promise<TransactionSignature> {
+    return await sendAndConfirmIxs(this.provider, await this.client.confirmOwnerTransferRequest());
+  }
+
+  async cancelOwnerTransferRequest(): Promise<TransactionSignature> {
+    return sendAndConfirmIxs(this.provider, await this.client.cancelOwnerTransferRequest());
+  }
+
+  async addAdmin(newAdmin: PublicKey): Promise<TransactionSignature> {
+    return sendAndConfirmIxs(this.provider, await this.client.addAdmin(newAdmin));
+  }
+
+  async removeAdmin(adminToRemove: PublicKey): Promise<TransactionSignature> {
+    return sendAndConfirmIxs(
+      this.provider,
+      await this.client.removeAdmin(this.publicKey, adminToRemove),
+    );
+  }
+
+  async registerPeer(chain: Chain, peerAddress: UniversalAddress): Promise<TransactionSignature> {
+    return sendAndConfirmIxs(
+      this.provider,
+      await this.client.registerPeer(this.publicKey, chain, peerAddress),
+    );
+  }
+
+  async updateCanonicalPeer(
+    chain: Chain,
+    peerAddress: UniversalAddress,
+  ): Promise<TransactionSignature> {
+    return sendAndConfirmIxs(
+      this.provider,
+      await this.client.updateCanonicalPeer(chain, peerAddress),
+    );
+  }
+
+  async setPauseForOutboundTransfers(chain: Chain, paused: boolean): Promise<TransactionSignature> {
+    return sendAndConfirmIxs(
+      this.provider,
+      await this.client.setPauseForOutboundTransfers(this.publicKey, chain, paused),
+    );
+  }
+
+  async updateMaxGasDropoff(chain: Chain, maxGasDropoff: number): Promise<TransactionSignature> {
+    return sendAndConfirmIxs(
+      this.provider,
+      await this.client.updateMaxGasDropoff(this.publicKey, chain, maxGasDropoff),
+    );
+  }
+
+  async updateRelayerFee(chain: Chain, relayerFee: number): Promise<TransactionSignature> {
+    return sendAndConfirmIxs(
+      this.provider,
+      await this.client.updateRelayerFee(this.publicKey, chain, relayerFee),
+    );
+  }
+
+  async updateFeeRecipient(newFeeRecipient: PublicKey): Promise<TransactionSignature> {
+    return sendAndConfirmIxs(
+      this.provider,
+      await this.client.updateFeeRecipient(this.publicKey, newFeeRecipient),
+    );
+  }
+
+  async updateEvmTransactionConfig(
+    evmTransactionGas: bigint,
+    evmTransactionSize: bigint,
+  ): Promise<TransactionSignature> {
+    return sendAndConfirmIxs(
+      this.provider,
+      await this.client.updateEvmTransactionConfig(
+        this.publicKey,
+        evmTransactionGas,
+        evmTransactionSize,
+      ),
+    );
+  }
+
+  async transferTokens(
+    params: TransferParameters,
+    signers?: Keypair[],
+  ): Promise<TransactionSignature> {
+    return sendAndConfirmIxs(
+      this.provider,
+      await this.client.transferTokens(this.publicKey, params),
+      { signers },
+    );
+  }
+
+  async completeTransfer(vaa: VaaMessage): Promise<TransactionSignature> {
+    return sendAndConfirmIxs(
+      this.provider,
+      await this.client.completeTransfer(this.publicKey, vaa),
+    );
+  }
+
+  async relayingFee(chain: Chain, dropoffAmount: number): Promise<number> {
+    return this.client.relayingFee(chain, dropoffAmount);
+  }
+}
